Create target backend for Nintendo Game Boy (DMG)
Create a new target backend for the Nintendo Game Boy (DMG) using GlobalISel.
This provides the minimal amount of plumbing to get the GISel instruction
selector to select a generic RET instruction, verified by an accompanying
test. This backend is not intended to be used with SelectionDAG at this point
in time, with the RET test reflecting this.

This provides a base to add support for more instructions incrementally,
eventually building up to a full backend for the Game Boy. To build llc to
target this backend, provide 'DMG' in the list of experimental targets when
configuring your local build.

Known issue: it is currently not possible to select the backend via the
-mtriple argument to llc. This is being investigated and will be fixed in a
subsequent commit.

Signed-off-by: Akram Ahmad <bulgeypip786@gmail.com>